Calculate FT height from ascent/descent
Since the `height` metric is not used by a lot of font rendering
systems, some fonts fail to properly define this metric. As a result
this can lead to clipped glyphs due to insufficient line heights.

This patch combines the ascent and descent from the baseline to
calculate a line height that should be reliably provided by every font.
The two metrics are then combined using whichever is bigger, to make
sure no glyphs are cut off accidentally.

This calculation is based on the behavior of other terminal emulators,
which most commonly use a combination of ascent and descent as the sole
source of font height (e.g. URxvt/Xterm/Konsole/VTE).