Remove duplicate Apple Symbols fallback
The normal Apple Symbols font is already included in the cascade list for Menlo, so there's no need to add it manually. I've removed that bit and rewritten the surrounding comments to be up to date and clearer.

A new test is also added, which will print a list of the fonts included in the Menlo cascade. This is just to make it easier to verify what exactly macOS is giving us.