Fix Core Text backend ignoring style for font fallback
Previously crossfont has always used Menlo-Regular as a original
fallback font for any font and included it in the fallback fonts. This
behavior means that crossfont ignores the font style, and fallback fonts
have always had regular style even if the current font has no regular
style.  Sometimes this behavior causes that CJK glyph have always been
rendered as regular even if they should be done as bold. The reasons are
below.

1. Causes font fallback frequently because many fonts have no CJK
   glyphs.
2. Lookups glyph at fallback fonts, but the fonts have only regular
   style.

We can solve this by changing the original fallback font style to the
current font style. The fallback font generator can return appropriate
style fonts according to the original fallback font style.

When we do it, we must choose which font is used as the original
fallback font. It was previously Menlo(-Regular) and was no problem, but
we don't know the reasonable reason to do it. Also probably no problem
too even if the original fallback font is the current font. But the
former is a little bit strange, especially in a way that the font name
is fixed in a source code. (If someone wants to use Menlo or another as
an original fallback font, maybe we should support the functionality to
do it as user configuration.) So use the latter. It is more natural in a
way that the fallback fonts are for the original fallback font and they
are generated using it.

We can see whether or not the style problem is solved using the command
below. It is done if all glyphs on the first line are regular and all
them on the second line are bold.

```bash
echo -e 'Hello, こんにちは, 你好, 안녕하세요\n\e[1mHello, こんにちは, 你好, 안녕하세요\e[0m'
```

Note that this fix should affect even if italic or bold italic for CJK
glyph in theory, but not in fact. Maybe no italic or bold italic CJK
fonts are installed by default in macOS. This is the same even if
terminal.app or iTerm2.