Add proportional font support
Since crossfont was previously used solely by Alacritty, there was no
need for anything other than monospaced fonts. This patch adds minimal
proportional font support to the FreeType backend.

The most relevant part for proportional fonts is using the glyph's
individual advance rather than the font metric's advance. This is
exposed as a new `advance` field on `RasterizedGlyph`.

Besides advance, it's also possible for two rasterized glyphs to be
moved closer together by the font in a process called kerning. Crossfont
mimics the FreeType API here and provides the `kerning` method which
returns the relative offset of the second of two glyphs. Kerning for
OpenType fonts implemented in a `GPOS` table is not supported.