Fix thin strokes on macOS
Remove the `use_thin_strokes` parameter from `Rasterize::new`, which
only did anything on macOS and only prior to Big Sur. Instead, we will
enable or disable "font smoothing" on macOS based on the
`AppleFontSmoothing` user default.

Subpixel anti-aliasing, which was previously controlled by the
`CGFontRenderingFontSmoothingDisabled` user default, is no longer
available, and the API which WebKit uses to read this setting now always
returns `true`, regardless of its value in the defaults dictionary. [1] [2]
The `set_font_smoothing` function of the `darwin` module, which set this
value in user defaults, has likewise been removed.

macOS itself supports values of `AppleFontSmoothing` from 0 to 3. If it
is unset the system behaves as though the value were 2. For `crossfont`,
we'll treat any value other than 0 to mean "smoothing enabled", as that
is the interface Core Graphics provides for this. A future improvement
may be to decipher what values of 1 and 3 mean for Core Graphics/Core
Text rendering.

These changes let users get the "thin strokes" behavior by setting
`AppleFontSmoothing` to 0 with:

```sh
$ defaults write -g AppleFontSmoothing -int 0
```

(Or replace `-g` with a specific domain the setting should apply to,
rather than the whole system.)

[1]: https://bugs.webkit.org/show_bug.cgi?id=216588
[2]: https://github.com/alacritty/crossfont/pull/47#issuecomment-1174723896

Fixes #36.